+++ include/render/cameras/ThinLensCamera.h
@@ -145,6 +145,7 @@ namespace render {
       *      used by tests.
       */
     Rayd rayForPixel(double x, double y, render::SampleStream& stream) const override;
+    std::unique_ptr<PrimaryRayGenerator> primaryRayGenerator() const override;
     std::shared_ptr<Camera> clone() const override;
     const char* fingerprintType() const override;
 
@@ -276,6 +277,8 @@ namespace render {
     void setViewPlane(std::shared_ptr<render::ViewPlane> plane) override;
 
   private:
+    Vector3d eyeOrigin() const;
+
     double m_distance;
     double m_zoom;
     double m_apertureRadius;